# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DentalAI SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dental/tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantBody"
      responses:
        "201":
          description: Created tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /dental/tenants/onboarding:
    post:
      operationId: onboardTenant
      tags: [tenants]
      summary: Onboard a new clinic (creates tenant + default settings + seed procedures)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardTenantBody"
      responses:
        "201":
          description: Onboarded tenant with settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardTenantResponse"
  /dental/tenants/{tenantId}:
    parameters:
      - $ref: "#/components/parameters/TenantIdParam"
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get a tenant by ID
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantBody"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete a tenant
      responses:
        "204":
          description: Deleted

  # ──────────────────────────── SUBSCRIPTION ────────────────────────────
...
components:
  parameters:
    TenantIdParam:
      name: tenantId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    OnboardTenantBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          default: free
        clinicName:
          type: string
          nullable: true
        clinicPhone:
          type: string
          nullable: true
        workingHoursStart:
          type: string
          default: "08:00"
        workingHoursEnd:
          type: string
          default: "18:00"
      required:
        - name
        - slug
    OnboardTenantResponse:
      type: object
      properties:
        tenant:
          $ref: "#/components/schemas/Tenant"
        settings:
          $ref: "#/components/schemas/DentalSettings"
        procedures:
          type: array
          items:
            $ref: "#/components/schemas/DentalProcedure"
      required:
        - tenant
        - settings
        - procedures
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
        subscriptionStatus:
          type: string
        subscribedAt:
          type: string
          format: date-time
          nullable: true
        subscriptionExpiresAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        whatsappProvider:
          type: string
        evolutionApiUrl:
          type: string
          nullable: true
        evolutionApiKey:
          type: string
          nullable: true
        evolutionInstanceName:
          type: string
          nullable: true
        uazapiHost:
          type: string
          nullable: true
        uazapiAdminToken:
          type: string
          nullable: true
        uazapiInstanceToken:
          type: string
          nullable: true
        uazapiInstanceId:
          type: string
          nullable: true
        elevenLabsApiKey:
          type: string
          nullable: true
        whatsappConnected:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - plan
        - subscriptionStatus
        - whatsappConnected
        - createdAt
        - updatedAt
...
    CreateTenantBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          default: free
      required:
        - name
        - slug
    UpdateTenantBody:
      type: object
      properties:
        name:
          type: string
        plan:
          type: string
        whatsappProvider:
          type: string
        evolutionApiUrl:
          type: string
          nullable: true
        evolutionApiKey:
          type: string
          nullable: true
        evolutionInstanceName:
          type: string
          nullable: true
        uazapiHost:
          type: string
          nullable: true
        uazapiAdminToken:
          type: string
          nullable: true
        uazapiInstanceToken:
          type: string
          nullable: true
        uazapiInstanceId:
          type: string
          nullable: true
        elevenLabsApiKey:
          type: string
          nullable: true
        openaiApiKey:
          type: string
          nullable: true

    # ─── PATIENT ───
...
    DentalProcedure:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: string
        durationMinutes:
          type: integer
        active:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - price
        - durationMinutes
        - active
        - createdAt
        - updatedAt
...
    DentalSettings:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        clinicName:
          type: string
          nullable: true
        clinicPhone:
          type: string
          nullable: true
        clinicAddress:
          type: string
          nullable: true
        professionalName:
          type: string
          nullable: true
        professionalGender:
          type: string
          nullable: true
          enum: ['male', 'female', 'other', null]
          description: "Dr. (male), Dra. (female), Dr(a). (other)"
        specialties:
          type: string
          nullable: true
        professionalCro:
          type: string
          nullable: true
        professionalSpecialties:
          type: string
          nullable: true
        workingHoursStart:
          type: string
        workingHoursEnd:
          type: string
        workingDays:
          type: string
        slotDurationMinutes:
          type: integer
        defaultLeadDurationMinutes:
          type: integer
        defaultPatientDurationMinutes:
          type: integer
        chargesConsultation:
          type: boolean
        consultationFee:
          type: string
        autoConfirmAppointments:
          type: boolean
        confirmDaysBefore:
          type: integer
        acceptsInsurance:
          type: boolean
        insurancePlans:
          type: string
          nullable: true
        insuranceDays:
          type: string
          nullable: true
        insuranceHoursStart:
          type: string
          nullable: true
        insuranceHoursEnd:
          type: string
          nullable: true
        lunchStart:
          type: string
          maxLength: 5
        lunchEnd:
          type: string
          maxLength: 5
        scheduleConfig:
          type: string
          nullable: true
        aiName:
          type: string
          maxLength: 100
        aiPersonality:
          type: string
          nullable: true
        personalityType:
          type: string
          nullable: true
        aiLanguage:
          type: string
        reminderHoursBefore:
          type: integer
        remarketingEnabled:
          type: boolean
        remarketingHours:
          type: string
        remarketingDays:
          type: string
        remarketingMaxLeads:
          type: integer
        remarketingIntervalHot:
          type: integer
        remarketingIntervalWarm:
          type: integer
        remarketingIntervalCold:
          type: integer
        followUpReminder:
          type: boolean
        followUpConfirmation:
          type: boolean
        followUpPostAppointment:
          type: boolean
        postAppointmentHoursAfter:
          type: integer
        noShowEnabled:
          type: boolean
        noShowPatientContactHoursAfter:
          type: integer
        noShowPatientMessage:
          type: string
          nullable: true
        birthdayEnabled:
          type: boolean
        birthdayHour:
          type: integer
          minimum: 0
          maximum: 23
        birthdayMessage:
          type: string
          nullable: true
        audioMode:
          type: string
          enum: [off, always, audio_reply_only]
        elevenLabsVoiceId:
          type: string
          nullable: true
        telegramBotToken:
          type: string
          nullable: true
        telegramChatId:
          type: string
          nullable: true
        telegramEscalationEnabled:
          type: boolean
        acceptsInstallments:
          type: boolean
          nullable: true
        maxInstallments:
          type: integer
          nullable: true
        acceptsBoleto:
          type: boolean
          nullable: true
        paymentNotes:
          type: string
          nullable: true
        recoveryEnabled:
          type: boolean
        recoveryInactivityDays:
          type: integer
        recoveryNoShowDays:
          type: integer
        recoveryAiInstructions:
          type: string
          nullable: true
        recoveryHours:
          type: string
        recoveryDays:
          type: string
        recoveryMaxPerRun:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - workingHoursStart
        - workingHoursEnd
        - workingDays
        - slotDurationMinutes
        - defaultLeadDurationMinutes
        - defaultPatientDurationMinutes
        - chargesConsultation
        - consultationFee
        - autoConfirmAppointments
        - confirmDaysBefore
        - acceptsInsurance
        - aiLanguage
        - reminderHoursBefore
        - audioMode
        - createdAt
        - updatedAt
...